Change namespace and strict typing
