+++ src/Core/Action/ActionsBarButtonsCollection.php
@@ -23,59 +23,19 @@
  * @copyright Since 2007 PrestaShop SA and Contributors
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
+declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Core\ActionBar;
+namespace PrestaShop\PrestaShop\Core\Action;
 
-class ActionsBarButton implements ActionsBarButtonInterface
+use PrestaShop\PrestaShop\Core\Data\AbstractTypedCollection;
+
+class ActionsBarButtonsCollection extends AbstractTypedCollection
 {
     /**
-     * @var string
+     * {@inheritdoc}
      */
-    protected $class;
-
-    /**
-     * @var string[]
-     */
-    protected $properties;
-
-    /**
-     * @var string
-     */
-    protected $content;
-
-    /**
-     * @param string $class
-     * @param string[] $properties
-     * @param string $content
-     */
-    public function __construct(string $class = '', array $properties = [], string $content = '')
+    protected function getType()
     {
-        $this->class = $class;
-        $this->properties = $properties;
-        $this->content = $content;
-    }
-
-    /**
-     * @return string
-     */
-    public function getClass(): string
-    {
-        return $this->class;
-    }
-
-    /**
-     * @return string[]
-     */
-    public function getProperties(): array
-    {
-        return $this->properties;
-    }
-
-    /**
-     * @return string
-     */
-    public function getContent(): string
-    {
-        return $this->content;
+        return ActionsBarButtonInterface::class;
     }
 }
